remove info module and move used functions from it to utils, refactor main module respectively
